requirements.txt
Flask==2.3.2
python-docx==0.8.11
language-tool-python==2.7.0
matplotlib==3.7.1
nltk==3.8.1
pandas==2.0.3
PyYAML==6.0
scikit-learn==1.3.0
yake==0.4.4
pyngrok==5.1.0
python-dotenv==1.0.0
pymorphy3==2.0.2